+++ Changes.md
@@ -1,4 +1,5 @@
 * fix missing output of Championship award in awards script
+* #978 - fix deleting of performance scores
 
 Release 16.1.1
 ==============